organize auth module
