+++ apps/api/v2/src/modules/auth/guards/next-auth/next-auth.guard.ts
