+++ apps/api/v2/src/modules/auth/guards/organization-roles/organization-roles.guard.ts
@@ -1,4 +1,4 @@
-import { Roles } from "@/modules/auth/decorator/roles/roles.decorator";
+import { Roles } from "@/modules/auth/decorators/roles/roles.decorator";
 import { MembershipRepository } from "@/modules/repositories/membership/membership.repository";
 import { Injectable, CanActivate, ExecutionContext } from "@nestjs/common";
 import { Reflector } from "@nestjs/core";
