+++ apps/api/v2/src/modules/auth/strategies/api-key-auth/api-key-auth.strategy.ts
