+++ apps/api/v2/src/modules/auth/strategies/index.ts
