+++ apps/api/v2/src/modules/auth/strategies/next-auth/next-auth.strategy.ts
